[BACKPORT 2025.2][#29249] docdb: (test-only) fix analyze test w/ vlogging enabled
Summary:
When the hardcoded constant `kEnableVlogsDuringAnalyze` is set to true, the `PgAnalyzeTest` tests restart the postmaster. The machinery they use to do this doesn't work after the introduction of the YSQL lease in 53146f9da4877244f931761920a5159a5b8da412.

This diff uses the tserver callback instead to restart pg, which works with the lease enabled.
Jira: DB-19024

Original commit: da0f5a6ab488d7137efa9ca783b3ca263afd2c19 / D48063

Test Plan:
apply diff:
```
-  static constexpr auto kEnableVlogsDuringAnalyze = false;
+  static constexpr auto kEnableVlogsDuringAnalyze = true;
```

run:
```
./yb_build.sh release --cxx-test pg_analyze-test --cxx-test-filter-re pg_analyze-test --gtest_filter PgAnalyzeTest.AnalyzeSamplingColocated --test-timeout-sec 1800
```

Reviewers: timur

Reviewed By: timur

Subscribers: ybase, rthallam, yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D48810

+++ src/yb/yql/pgwrapper/pg_mini_test_base.h
@@ -30,6 +30,7 @@ namespace yb::pgwrapper {
 
 class PgMiniTestBase : public MiniClusterTestWithClient<MiniCluster> {
  protected:
+  // Use TS-0 IP for PG server. YBC process and PG auto analyze service use this IP.
   constexpr static size_t kPgTsIndex = 0;
 
   // This allows modifying flags before we start the postgres process in SetUp.
@@ -85,7 +86,7 @@ class PgMiniTestBase : public MiniClusterTestWithClient<MiniCluster> {
 
   virtual void StartPgSupervisor(uint16_t pg_port, const int pg_ts_idx);
 
-  Status SetupPGCallbacksAndStartPG(uint16_t pg_port, int pg_ts_idx, bool wait_for_pg = true);
+  Status SetupPGCallbacksAndStartPG(uint16_t pg_port, size_t pg_ts_idx, bool wait_for_pg = true);
 
   Status RestartPostgres();
